completed story-7 Pricing strategy according to number of seats

+++ src/main/java/com/everest/airline/data/DataReader.java
@@ -36,9 +36,9 @@ public class DataReader {
             int occupiedEconomicSeats = Integer.parseInt(flight[7]);
             int occupiedFirstClassSeats = Integer.parseInt(flight[8]);
             int occupiedSecondClassSeats = Integer.parseInt(flight[9]);
-            int economyFare = Integer.parseInt(flight[10]);
-            int firstClassFare = Integer.parseInt(flight[11]);
-            int secondClassFare = Integer.parseInt(flight[12]);
+            double economyFare = Double.parseDouble(flight[10]);
+            double firstClassFare = Double.parseDouble(flight[11]);
+            double secondClassFare = Double.parseDouble(flight[12]);
             flightsFromFile.add(new Flight(flightNumber, source, destination, LocalDate.of(year, month, day), economicSeatCapacity,FirstClassSeatCapacity,secondClassSeatCapacity,occupiedEconomicSeats,occupiedFirstClassSeats,occupiedSecondClassSeats,economyFare,firstClassFare,secondClassFare));
         }
         return flightsFromFile;
